+++ src/theme/index.ts
@@ -20,6 +20,7 @@ import { CustomImage, CustomImageWrapper } from "./components/image";
 import CustomSelect from "./components/select";
 import DatePicker from "./components/datePicker";
 import Fieldset from "./components/fieldset";
+import Footer from "./components/footer";
 import Heading from "./components/heading";
 import HelperErrorText from "./components/helperErrorText";
 import Hero from "./components/hero";
@@ -92,6 +93,7 @@ const theme = extendTheme({
     CustomSelect,
     DatePicker,
     Fieldset,
+    Footer,
     Heading,
     HelperErrorText,
     Hero,