+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the basic `Footer` React component.
+
 ## 1.0.5 (July 7, 2022)
 
 ### Adds